Add files via upload
This gives us the RA & DEC values for an asteroid, using PyOorb, given its orbital parameters. I've used Ceres as an example, but I also made a function (named "getOrbitAtTime") that generalizes this and returns the output as an orbit. I'm currently working on getting the RA and DEC from that data (which I think should each be made into functions for ease of readability - e.g. "getRA(orbit)" and "getDEC(orbit)"